Fixed AOSP style and added HostnameVerifierTest
