+++ testing/src/main/java/org/conscrypt/javax/net/ssl/FakeSSLSession.java
@@ -19,6 +19,7 @@ package org.conscrypt.javax.net.ssl;
 import java.nio.charset.Charset;
 import java.security.Principal;
 import java.security.cert.Certificate;
+import javax.net.ssl.SSLPeerUnverifiedException;
 import javax.net.ssl.SSLSession;
 import javax.net.ssl.SSLSessionContext;
 
@@ -76,7 +77,7 @@ public class FakeSSLSession implements SSLSession {
     }
 
     @Override
-    public Certificate[] getPeerCertificates() {
+    public Certificate[] getPeerCertificates() throws SSLPeerUnverifiedException {
         throw new UnsupportedOperationException();
     }
